requirements.txt
tensorflow>=2.10.0
streamlit
pillow
pandas
numpy
matplotlib